+++ programs/ft_memcpy.c
@@ -6,13 +6,13 @@
 /*   By: abosc <abosc@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/15 10:28:20 by abosc             #+#    #+#             */
-/*   Updated: 2024/10/15 10:55:27 by abosc            ###   ########.fr       */
+/*   Updated: 2024/10/15 12:44:58 by abosc            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <string.h>
+// #include <string.h>
 
-void	*memcpy(void *dest, const void *src, size_t n)
+void	*memcpy(void *dest, const void *src, unsigned int n)
 {
 	int				i;
 	unsigned char	*modifiable_src;
